remove "Respond"
The "Respond" activity type was added speculatively as a
superclass of a set of more specific Activity types like
"Like", "Dislike", etc. Since it was added, the decision
was made not to require support for inferencing, making
it largely unnecessary to define a hierarchy of responses.
None of our user stories justify having a separate generic
"Respond" activity and we already have "in-reply-to" as
a property.