Refactor as:Connection
Rename as:Connection to as:Relationship
Rename as:a to as:subject
Remove as:b, reuse as:object instead.

As a side effect of this change, a single Relationship
object can be used to indicate multiple relationships.
as:subject is functional, but as:object is not.

e.g. "Sally is friends with Joe and Karen" :

```
{
  "@type": "Relationship",
  "subject": "http://sally.example.org",
  "relationship": "http://example.org/friendsWith",
  "object": [
    "http://joe.example.org",
    "http://karen.example.org"
  ]
}
```